Moved previous lib.rs implementation to egfx.rs
